+++ bullet-featherstone/src/SDFFeatures.cc
@@ -500,7 +500,7 @@ Identity SDFFeatures::ConstructSdfModelImpl(
 
   std::unordered_map<const ::sdf::Model*, Identity> modelIDs;
   std::size_t rootModelID = 0u;
-  std::shared_ptr<btMultiBody> rootMultiBody;
+  std::shared_ptr<GzMultiBody> rootMultiBody;
   // Add all  models, including nested models
   auto addModels = [&](std::size_t _modelOrWorldID, const ::sdf::Model *_model,
                        auto &&_addModels)
@@ -525,7 +525,7 @@ Identity SDFFeatures::ConstructSdfModelImpl(
         {
           auto worldIdentity = this->GenerateIdentity(
               _modelOrWorldID, worldIt->second);
-          rootMultiBody = std::make_shared<btMultiBody>(
+          rootMultiBody = std::make_shared<GzMultiBody>(
                 static_cast<int>(structure.flatLinks.size()),
                 structure.mass,
                 structure.inertia,
@@ -785,8 +785,6 @@ Identity SDFFeatures::ConstructSdfModelImpl(
             static_cast<btScalar>(joint->Axis()->Lower());
         model->body->getLink(i).m_jointUpperLimit =
             static_cast<btScalar>(joint->Axis()->Upper());
-        model->body->getLink(i).m_jointDamping =
-            static_cast<btScalar>(joint->Axis()->Damping());
         model->body->getLink(i).m_jointFriction =
             static_cast<btScalar>(joint->Axis()->Friction());
         model->body->getLink(i).m_jointMaxVelocity =
@@ -800,6 +798,7 @@ Identity SDFFeatures::ConstructSdfModelImpl(
         jointInfo->maxVelocity = joint->Axis()->MaxVelocity();
         jointInfo->axisLower = joint->Axis()->Lower();
         jointInfo->axisUpper = joint->Axis()->Upper();
+        jointInfo->damping = joint->Axis()->Damping();
 
         jointInfo->jointLimits =
           std::make_shared<btMultiBodyJointLimitConstraint>(
@@ -872,7 +871,7 @@ Identity SDFFeatures::ConstructSdfModelImpl(
     *worldToModel * modelToNestedModel * *modelToRootLink *
     rootInertialToLink.inverse();
 
-  model->body->setBaseWorldTransform(convertTf(worldToRootCom));
+  model->body->SetBaseWorldTransform(convertTf(worldToRootCom));
   model->body->setBaseVel(btVector3(0, 0, 0));
   model->body->setBaseOmega(btVector3(0, 0, 0));
 
@@ -1484,11 +1483,12 @@ void SDFFeatures::CreateLinkCollider(const Identity &_linkID, bool _isStatic,
     // check if it's a base link
     isFixed = std::size_t(_linkID) ==
         static_cast<std::size_t>(modelInfo->body->getUserIndex());
-    // check if link has zero dofs
+    // check if link has zero dofs from model base.
     if (!isFixed && linkInfo->indexInModel.has_value())
     {
-       isFixed = modelInfo->body->getLink(
-          linkInfo->indexInModel.value()).m_dofCount == 0;
+      auto link = modelInfo->body->getLink(linkInfo->indexInModel.value());
+      int totalLinkDofs = link.m_dofOffset + link.m_dofCount;
+      isFixed = totalLinkDofs == 0;
     }
   }
   if (_isStatic || isFixed)