+++ dartsim/src/SDFFeatures_TEST.cc
@@ -52,7 +52,7 @@
 #include <sdf/World.hh>
 
 #include <test/Utils.hh>
-#include <test/common_test/Worlds.hh>
+#include "test/common_test/Worlds.hh"
 #include "Worlds.hh"
 
 #include "World.hh"